from int to float for position things

+++ Engine/Cube.cpp
@@ -2,7 +2,7 @@
 
 void Cube::Draw(Graphics& gfx)
 {
-	gfx.DrawRectColorDim(x, y, width, height, color);
+	gfx.DrawRectColorDim(int(x), int(y), width, height, color);
 }
 
 bool Cube::IsEaten() const
@@ -54,19 +54,19 @@ void Cube::UpdateColorChili()
 
 void Cube::TestCollide(const Dude & dude)
 {
-	const int box1Left = x;
-	const int box1Right = x + width - 1;
-	const int box1Up = y;
-	const int box1Down = y + height - 1;
-	const int box2Left = dude.GetX();
-	const int box2Right = dude.GetX() + Dude::width - 1;
-	const int box2Up = dude.GetY();
-	const int box2Down = dude.GetY() + Dude::height - 1;
+	const float box1Left = x;
+	const float box1Right = x + float(width - 1);
+	const float box1Up = y;
+	const float box1Down = y + float(height - 1);
+	const float box2Left = dude.GetX();
+	const float box2Right = dude.GetX() + float(Dude::width - 1);
+	const float box2Up = dude.GetY();
+	const float box2Down = dude.GetY() + float(Dude::height - 1);
 	isEaten = box1Left <= box2Right && box1Right >= box2Left &&
 		box1Up <= box2Down && box1Down >= box2Up;
 }
 
-void Cube::InitXY(int x_1, int y_1)
+void Cube::InitXY(float x_1, float y_1)
 {
 	x = x_1;
 	y = y_1;
